Moved SqlParser class to parser package
